[API] Move attachOrganizations to target-profile-share-repository
