+++ api/lib/domain/usecases/attach-organizations-to-target-profile.js
@@ -3,11 +3,11 @@ const { TargetProfileOrganizations } = require('../models');
 module.exports = async function attachOrganizationsToTargetProfile({
   targetProfileId,
   organizationIds,
-  targetProfileRepository,
+  targetProfileShareRepository,
 }) {
   const targetProfileOrganizations = new TargetProfileOrganizations({ id: targetProfileId });
 
   targetProfileOrganizations.attach(organizationIds);
 
-  return targetProfileRepository.attachOrganizations(targetProfileOrganizations);
+  return targetProfileShareRepository.attachOrganizations(targetProfileOrganizations);
 };
