+++ api/tests/integration/infrastructure/repositories/target-profile-repository_test.js
@@ -1,5 +1,5 @@
 const _ = require('lodash');
-const { expect, databaseBuilder, domainBuilder, catchErr, sinon, knex } = require('../../../test-helper');
+const { expect, databaseBuilder, catchErr, sinon, knex } = require('../../../test-helper');
 const TargetProfile = require('../../../../lib/domain/models/TargetProfile');
 const TargetProfileForCreation = require('../../../../lib/domain/models/TargetProfileForCreation');
 const Skill = require('../../../../lib/domain/models/Skill');
@@ -486,91 +486,6 @@ describe('Integration | Repository | Target-profile', function () {
 
         // then
         expect(_.map(matchingTargetProfiles, 'id')).to.have.members([targetProfileId1, targetProfileId2]);
-      });
-    });
-  });
-
-  describe('#attachOrganizations', function () {
-    afterEach(function () {
-      return knex('target-profile-shares').delete();
-    });
-
-    it('should return attachedIds', async function () {
-      const targetProfileId = databaseBuilder.factory.buildTargetProfile().id;
-      const organization1 = databaseBuilder.factory.buildOrganization();
-      const organization2 = databaseBuilder.factory.buildOrganization();
-
-      await databaseBuilder.commit();
-
-      const targetProfileOrganizations = domainBuilder.buildTargetProfileOrganizations({ id: targetProfileId });
-
-      targetProfileOrganizations.attach([organization1.id, organization2.id]);
-
-      const results = await targetProfileRepository.attachOrganizations(targetProfileOrganizations);
-
-      expect(results).to.deep.equal({ duplicatedIds: [], attachedIds: [organization1.id, organization2.id] });
-    });
-
-    it('add organization to the target profile', async function () {
-      const targetProfileId = databaseBuilder.factory.buildTargetProfile().id;
-      const organization1 = databaseBuilder.factory.buildOrganization();
-      const organization2 = databaseBuilder.factory.buildOrganization();
-
-      await databaseBuilder.commit();
-
-      const targetProfileOrganizations = domainBuilder.buildTargetProfileOrganizations({ id: targetProfileId });
-
-      targetProfileOrganizations.attach([organization1.id, organization2.id]);
-
-      await targetProfileRepository.attachOrganizations(targetProfileOrganizations);
-
-      const rows = await knex('target-profile-shares')
-        .select('organizationId')
-        .where({ targetProfileId: targetProfileOrganizations.id });
-      const organizationIds = rows.map(({ organizationId }) => organizationId);
-
-      expect(organizationIds).to.exactlyContain([organization1.id, organization2.id]);
-    });
-
-    context('when the organization does not exist', function () {
-      it('throws an error', async function () {
-        const targetProfileId = databaseBuilder.factory.buildTargetProfile().id;
-        const organizationId = databaseBuilder.factory.buildOrganization().id;
-        const unknownOrganizationId = 99999;
-
-        await databaseBuilder.commit();
-
-        const targetProfileOrganizations = domainBuilder.buildTargetProfileOrganizations({ id: targetProfileId });
-
-        targetProfileOrganizations.attach([unknownOrganizationId, organizationId]);
-
-        const error = await catchErr(targetProfileRepository.attachOrganizations)(targetProfileOrganizations);
-
-        expect(error).to.be.an.instanceOf(NotFoundError);
-        expect(error.message).to.have.string(`L'organization avec l'id ${unknownOrganizationId} n'existe pas`);
-      });
-    });
-
-    context('when the organization is already attached', function () {
-      it('should return inserted organizationId', async function () {
-        const targetProfileId = databaseBuilder.factory.buildTargetProfile().id;
-        const firstOrganization = databaseBuilder.factory.buildOrganization();
-        const secondOrganization = databaseBuilder.factory.buildOrganization();
-
-        databaseBuilder.factory.buildTargetProfileShare({
-          targetProfileId: targetProfileId,
-          organizationId: firstOrganization.id,
-        });
-
-        await databaseBuilder.commit();
-
-        const targetProfileOrganizations = domainBuilder.buildTargetProfileOrganizations({ id: targetProfileId });
-
-        targetProfileOrganizations.attach([firstOrganization.id, secondOrganization.id]);
-
-        const result = await targetProfileRepository.attachOrganizations(targetProfileOrganizations);
-
-        expect(result).to.deep.equal({ duplicatedIds: [firstOrganization.id], attachedIds: [secondOrganization.id] });
       });
     });
   });
